// src/app/models/Product.ts
export interface Product {
    id: number;
    productName: string;
    price: number;
    priceHasDecreased: number;
    description: string;
    quantity: number;
    image: string;
    create: string;
    brandId: number;
    brandName: string;
    typeId: number;
    typeName: string;
    colors: string;
    status: number;
}